requirements.txt
# Core async
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.25.0
pydantic>=2.5.0

# Search
rapidfuzz>=3.6.0

# Image processing / conversion
Pillow>=10.2.0

# Progress bars
tqdm>=4.66.0

# Config
PyYAML>=6.0.1

# HTTP (for legacy/fallback)
requests>=2.31.0
